move responsibility for appending entry footer
from loader to selector
