+++ linen-repository/src/main/scala/x7c1/linen/repository/entry/unread/EntryAccessor.scala
@@ -8,9 +8,6 @@ import x7c1.linen.repository.unread.{EntryKind, SourceKind, UnreadRowKind}
 import x7c1.wheat.macros.database.TypedCursor
 import x7c1.wheat.modern.sequence.Sequence
 
-import scala.annotation.tailrec
-import scala.collection.mutable.ListBuffer
-
 trait EntryAccessor[+A <: UnreadEntry] extends Sequence[EntryRowContent[A]] {
 
   def latestEntriesTo(position: Int): Seq[A]
@@ -24,110 +21,6 @@ trait EntryAccessor[+A <: UnreadEntry] extends Sequence[EntryRowContent[A]] {
 
 trait ClosableEntryAccessor[+A <: UnreadEntry] extends EntryAccessor[A]{
   def close(): Unit
-}
-
-class EntryRowsBinder[A <: UnreadEntry] private (
-  accessors: ListBuffer[ClosableEntryAccessor[A]]) extends ClosableEntryAccessor[A]{
-
-  def append(sequence: ClosableEntryAccessor[A]): Unit = {
-    accessors += sequence
-  }
-  override def findAt(position: Int) = {
-    findAccessor(accessors, position, 0) flatMap { case (accessor, prev) =>
-      accessor.findAt(position - prev)
-    }
-  }
-
-  override def length: Int = {
-    accessors.foldLeft(0){ _ + _.length }
-  }
-
-  override def firstEntryPositionOf(sourceId: Long): Option[Int] = {
-    @tailrec
-    def loop(accessors: Seq[EntryAccessor[A]], prev: Int): Option[Int] =
-      accessors match {
-        case x +: xs => x.firstEntryPositionOf(sourceId) match {
-          case Some(s) => Some(prev + s)
-          case None => loop(xs, x.length + prev)
-        }
-        case Seq() => None
-      }
-
-    loop(accessors, 0)
-  }
-
-  override def findKindAt(position: Int) = {
-    findAccessor(accessors, position, 0) flatMap { case (accessor, prev) =>
-      accessor.findKindAt(position - prev)
-    }
-  }
-
-  @tailrec
-  private def findAccessor(
-    accessors: Seq[EntryAccessor[A]],
-    position: Int,
-    prev: Int): Option[(EntryAccessor[A], Int)] = {
-
-    accessors match {
-      case x +: xs => x.length + prev match {
-        case sum if sum > position => Some(x -> prev)
-        case sum => findAccessor(xs, position, sum)
-      }
-      case Seq() => None
-    }
-  }
-
-  override def close(): Unit = synchronized {
-    accessors foreach (_.close())
-    accessors.clear()
-  }
-  override def lastEntriesTo(position: Int) = {
-    @tailrec
-    def loop(
-      accessors: Seq[EntryAccessor[A]],
-      remains: Int, entries: Seq[A]): Seq[A] = {
-
-      accessors match {
-        case head +: tail =>
-          val xs = entries ++ head.lastEntriesTo(remains)
-          val diff = remains - head.length
-          if (diff <= 0){
-            xs
-          } else {
-            loop(tail, diff, xs)
-          }
-        case Seq() => entries
-      }
-    }
-    loop(accessors, position, Seq())
-  }
-
-  override def latestEntriesTo(position: Int): Seq[A] = {
-    @tailrec
-    def loop(
-      accessors: Seq[EntryAccessor[A]],
-      remains: Int, entries: Seq[A]): Seq[A] = {
-
-      accessors match {
-        case head +: tail =>
-          val xs = entries ++ head.latestEntriesTo(remains)
-          val diff = remains - head.length
-          if (diff <= 0){
-            xs
-          } else {
-            loop(tail, diff, xs)
-          }
-        case Seq() => entries
-      }
-    }
-    loop(accessors, position, Seq())
-  }
-}
-
-object EntryRowsBinder {
-  def apply[A <: UnreadEntry](): EntryRowsBinder[A] = {
-    new EntryRowsBinder[A](new ListBuffer())
-  }
 }
 
 class EntryAccessorImpl[A <: UnreadEntry](
